Sending Packets and a lot of magic

+++ RouteMeMobile/src/de/unipotsdam/nexplorer/client/android/sensors/ShakeDetector.java
@@ -72,7 +72,7 @@ public class ShakeDetector implements SensorEventListener {
 		float y = se.values[1];
 		float z = se.values[2];
 		mAccelLast = mAccelCurrent;
-		mAccelCurrent = (float) Math.sqrt((double) (x * x + y * y + z * z));
+		mAccelCurrent = (float) Math.sqrt(x * x + y * y + z * z);
 		float delta = mAccelCurrent - mAccelLast;
 		mAccel = mAccel * 0.9f + delta * 0.1f; // perform low-cut filter
